Changed RDMs test in DMET to test the traces.
