+++ qsdk/problem_decomposition/tests/dmet/test_dmet.py
@@ -19,6 +19,8 @@ from qsdk.molecule_library import mol_H4_doublecation_minao, mol_H4_doublecation
 from qsdk.problem_decomposition import dmet
 from qsdk.problem_decomposition.dmet.dmet_problem_decomposition import Localization, DMETProblemDecomposition
 from qsdk.algorithms.variational import VQESolver
+from qsdk.toolboxes.molecular_computation.rdms import matricize_2rdm
+
 
 class DMETProblemDecompositionTest(unittest.TestCase):
 
@@ -211,7 +213,7 @@ class DMETProblemDecompositionTest(unittest.TestCase):
                         [[ 2.53371393e-07, -2.03773912e-05],
                         [-2.03773912e-05,  1.63885145e-03]]]]
 
-        fragment, q_H, q_circuit = dmet_solver.quantum_fragments_data[0]
+        fragment, _, q_circuit = dmet_solver.quantum_fragments_data[0]
 
         vqe_solver = VQESolver({"molecule": fragment, "ansatz": q_circuit,
                                 "qubit_mapping": "scBK"})
@@ -220,8 +222,15 @@ class DMETProblemDecompositionTest(unittest.TestCase):
 
         onerdm, twordm = vqe_solver.get_rdm(vqe_solver.optimal_var_params)
 
-        np.testing.assert_array_almost_equal(ref_onerdm, onerdm, decimal=3)
-        np.testing.assert_array_almost_equal(ref_twordm, twordm, decimal=3)
+        # Test traces of matrices
+        n_elec = fragment.n_active_electrons
+        n_orb = fragment.n_active_sos // 2
+        self.assertAlmostEqual(np.trace(onerdm), n_elec, delta=1e-3,
+                               msg="Trace of one_rdm does not match number of electrons")
+
+        rho = matricize_2rdm(twordm, n_orb)
+        self.assertAlmostEqual(np.trace(rho), n_elec * (n_elec - 1), delta=1e-3,
+                               msg="Trace of two_rdm does not match n_elec * (n_elec-1)")
 
 
 if __name__ == "__main__":
