+++ qsdk/toolboxes/molecular_computation/rdms.py
@@ -1,0 +1,39 @@
+# Copyright 2021 1QB Information Technologies Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Module containing functions to manipulate 1- and 2-RDMs."""
+
+import numpy as np
+
+
+def matricize_2rdm(two_rdm, n_orbitals):
+    """Turns the two_rdm tensor into a matrix for test purposes."""
+
+    l = 0
+    sq = n_orbitals * n_orbitals
+    jpqrs = np.zeros((n_orbitals, n_orbitals), dtype=int)
+    for i in range(n_orbitals):
+        for j in range(n_orbitals):
+            jpqrs[i, j] = l
+            l += 1
+
+    rho = np.zeros((sq, sq), dtype=complex)
+    for i in range(n_orbitals):
+        for j in range(n_orbitals):
+            ij = jpqrs[i, j]
+            for k in range(n_orbitals):
+                for l in range(n_orbitals):
+                    kl = jpqrs[k, l]
+                    rho[ij, kl] += two_rdm[i, k, j, l]
+    return rho
